+++ lib/widgets/bookshelf/book_bottom_sheet.dart
@@ -21,6 +21,7 @@ import 'package:anx_reader/widgets/bookshelf/book_cover.dart';
 import 'package:anx_reader/widgets/delete_confirm.dart';
 import 'package:anx_reader/widgets/icon_and_text.dart';
 import 'package:file_picker/file_picker.dart';
+import 'package:anx_reader/widgets/common/anx_button.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_smart_dialog/flutter_smart_dialog.dart';
@@ -102,13 +103,13 @@ class BookBottomSheet extends ConsumerWidget {
               ],
             ),
             actions: [
-              TextButton(
+              AnxButton.text(
                 onPressed: () {
                   SmartDialog.dismiss();
                 },
                 child: Text(L10n.of(context).commonCancel),
               ),
-              TextButton(
+              AnxButton.text(
                 onPressed: () {
                   SmartDialog.dismiss();
                   core();
@@ -182,13 +183,13 @@ class BookBottomSheet extends ConsumerWidget {
             ],
           ),
           actions: [
-            TextButton(
+            AnxButton.text(
               onPressed: () {
                 SmartDialog.dismiss(result: false);
               },
               child: Text(L10n.of(context).commonCancel),
             ),
-            TextButton(
+            AnxButton.text(
               onPressed: () {
                 SmartDialog.dismiss(result: true);
               },